Fixed an integer overflow that occurs due to a large continuum size

+++ de.tudarmstadt.ukp.statistics/src/main/java/de/tudarmstadt/ukp/dkpro/statistics/unitizing/AlphaUnitizedAgreement.java
@@ -153,8 +153,8 @@ public class AlphaUnitizedAgreement
 
         // calculate denominator
         double denominator = 0.0;
-        denominator = (study.getAnnotatorCount() * study.getContinuumLength())
-                * (study.getAnnotatorCount() * study.getContinuumLength() - 1);
+        denominator = ((double)study.getAnnotatorCount() * (double)study.getContinuumLength())
+                * ((double)study.getAnnotatorCount() * (double)study.getContinuumLength() - 1);
         double tmp = 0.0;
         for (int i = 0; i < study.getAnnotatorCount(); i++) {
             final List<Section> sections = study.getSections(category, i);
